refactor: rename and move error types
