+++ frontend/svelte/src/lib/types/ledger.errors.ts
