+++ frontend/svelte/src/lib/utils/ledger.utils.ts
@@ -8,7 +8,7 @@ import {
   LEDGER_SIGNATURE_LENGTH,
   type AllLedgerError,
 } from "../constants/ledger.constants";
-import { LedgerErrorKey, LedgerErrorMessage } from "../errors/ledger.errors";
+import { LedgerErrorKey, LedgerErrorMessage } from "../types/ledger.errors";
 import { Secp256k1PublicKey } from "../keys/secp256k1";
 import { i18n } from "../stores/i18n";
 import { replacePlaceholders } from "./i18n.utils";
